Put the sensor reporting into a loop
Put the logger script into a loop that will report once a minute for an hour, added the start of some try / except error handling which will prevent the script from crashing on an error. 

To do: 
Write the error handling logic, report to the errors feed if possible and find alternate reporting method for errors like 403 which are not network failures.