Create test.py
Added a test script to run for troubleshooting purposes and prior to installing the systemd service. 

Confirms the sensor is working correctly by printing the temp and humidity data to the console.
Send data to the Adafruit feeds immediately rather than waiting for the service to start 5 minutes after it is enabled.