Added a step for testing the sensor to script
Used the new test.py file included in the repo as a test of the sensor and Adafruit connection prior to the last step of installing the systemd service. 

This should allow users to immediately confirm the sensor is working as expected and check the feeds that they've configured on Adafruit to confirm data is sent to them correctly.